recommendation-service/requirements.txt
# Python Recommendation Engine Dependencies
# Python 3.8+ required

# Core ML Libraries
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# Database & Cache
mysql-connector-python==8.2.0
redis==5.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Optional: Advanced ML (uncomment if needed)
# scipy==1.11.4
# implicit==0.7.2  # For collaborative filtering
